// school-frontend/src/api/school.ts
export interface CreateStudentRequest {
  first_name: string;
  last_name: string;
  email: string;
}
// ...
export interface Student {
  student_id?: number | null;
  first_name: string;
  last_name: string;
  email: string;
}